avoid unnecessary casting

+++ bitarray/_util.c
@@ -127,7 +127,7 @@ count_to_n(bitarrayobject *a, Py_ssize_t n)
     while (i + BLOCK_BITS < a->nbits) {
         m = 0;
         assert(i % 8 == 0);
-        block_start = (Py_ssize_t) (i / 8);
+        block_start = i / 8;
         block_stop = block_start + (BLOCK_BITS / 8);
         for (k = block_start; k < block_stop; k++) {
             assert(k < Py_SIZE(a));
@@ -142,7 +142,7 @@ count_to_n(bitarrayobject *a, Py_ssize_t n)
 #undef BLOCK_SIZE
 
     while (i + 8 < a->nbits) {
-        k = (Py_ssize_t) (i / 8);
+        k = i / 8;
         assert(k < Py_SIZE(a));
         c = a->ob_item[k];
         m = bitcount_lookup[c];
